Update axi4lite_pb_bridge to v1.7.
v1.7 pipelines the port access to make it easier to meet timing.
That is, the port access is decoded in the cycle before write_strobe
and k_write_strobe go. Because the port accesses are different
for write_strobe and k_write_strobe (you check the top nybble for
write_strobe, but do not for k_write_strobe), the module was changed
to accept k_write_strobe directly.

If k_write_strobe is not used in this module, just tie it to 0.
Do not pass k_write_strobe if someone else uses it!

Note that you *can* still use the constant space even if the base
address has to be non-zero for some other reason. k_write_strobe
only checks the low nybble. So if you set BASE_ADDRESS to 20h,
you can write variable data to 20h, and constant data to 00h, and
another module can write variable data to 00h. (I don't know why
you would do this, but to each his own).

Note that the intermediate versions here didn't work (they *almost* did...)
hence the large version skip.